fix: break some circular imports
